Automatsko uzimanje trenutne lokacije

+++ app/src/main/java/com/pma/weatherapp/base/model/air_pollution/Components.kt
@@ -2,7 +2,7 @@ package com.pma.weatherapp.base.model.air_pollution
 
 data class Components(
     val co: Double,
-    val no: Int,
+    val no: Double,
     val no2: Double,
     val o3: Double,
     val so2: Double,
